influencehub/backend/requirements.txt

# ============================================
# backend/requirements.txt
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
mysql-connector-python==9.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1